requirements.txt
attrs==19.1.0
Django==2.2.4
djangorestframework==3.10.2
jsonschema==3.0.2
mysqlclient==1.4.4
numpy==1.17.0
protobuf==3.9.1
pyrsistent==0.15.4
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
